Split RuleResults from CourseList
